+++ packages/cubejs-testing-drivers/src/helpers/index.ts
@@ -4,6 +4,7 @@ import { getTempPath } from './getTempPath';
 import { getComposePath } from './getComposePath';
 import { getCreateQueries } from './getCreateQueries';
 import { getSelectQueries } from './getSelectQueries';
+import { getRefreshQueries } from './getRefreshQueries';
 import { getCubeJsPath } from './getCubeJsPath';
 import { getFixtures } from './getFixtures';
 import { getSchemaPath } from './getSchemaPath';
@@ -21,6 +22,7 @@ export {
   getComposePath,
   getCreateQueries,
   getSelectQueries,
+  getRefreshQueries,
   getCubeJsPath,
   getFixtures,
   getSchemaPath,